Linked cards from homepage to direct post page.
Removed date in props.

Card styles.
